[RAP] Fix UT compile: Datum(long long) has no variant alternative (rap_index_test.cpp:3088, :3181)
`Datum::Variant` holds `long` (int64_t on this target) but not `long long`, so
`Datum(1785058200000LL + r)` -- whose operands promote to `long long` -- fails to
instantiate `Datum::set<long long>`:

  be/src/types/datum.h:149:20: error: no match for 'operator=' (operand types are
  'starrocks::Datum::Variant' ... and 'const long long int')

Both sites wrapped in `static_cast<int64_t>(...)`. Same 64-bit two's-complement
value, no assertion or expected value changed. Blocks every starrocks_test suite.

+++ be/test/formats/parquet/rap_index_test.cpp
@@ -3085,7 +3085,7 @@ TEST_F(RapIndexTest, V5BuilderChoosesShapeOnExactBytes) {
     formats::RapSidecarBuilder uniq("event_time", 2, TYPE_BIGINT, 20000);
     {
         auto col = ColumnHelper::create_column(TypeDescriptor::from_logical_type(TYPE_BIGINT), false);
-        for (int64_t r = 0; r < 100000; ++r) col->append_datum(Datum(1785058200000LL + r));
+        for (int64_t r = 0; r < 100000; ++r) col->append_datum(Datum(static_cast<int64_t>(1785058200000LL + r)));
         uniq.observe(*col, 0);
     }
     EXPECT_EQ(uniq.num_values(), 100000u);
@@ -3178,7 +3178,7 @@ TEST_F(RapIndexTest, V5BytesMatchTheReferenceEncoder) {
         auto col = ColumnHelper::create_column(TypeDescriptor::from_logical_type(TYPE_BIGINT), true);
         for (int64_t r = 0; r < 100000; ++r) {
             if (r == 99999) col->append_nulls(1);
-            else col->append_datum(Datum(1785058200000LL + (r / 20000) * 10 + (r % 10)));
+            else col->append_datum(Datum(static_cast<int64_t>(1785058200000LL + (r / 20000) * 10 + (r % 10))));
         }
         uniq.observe(*col, 0);
     }
